+++ lib/threatinator/actions/list/action.rb
@@ -1,6 +1,7 @@
 require 'threatinator/action'
 require 'threatinator/exceptions'
 require 'csv'
+require 'multi_json'
 
 module Threatinator
   module Actions
@@ -11,8 +12,7 @@ module Threatinator
           @config = config
         end
 
-        def exec
-          io_out = $stdout
+        def output_table(io_out)
           feed_info = [['provider', 'name', 'type', 'link/path']]
           registry.each do |feed|
             info = [ feed.provider, feed.name ]
@@ -46,6 +46,40 @@ module Threatinator
           end
           io_out.printf(fmt, *sep)
           io_out.puts("Total: #{feed_info.count}")
+        end
+
+        def output_json(io_out)
+          feeds = []
+          registry.each do |feed|
+            info = {
+              provider: feed.provider,
+              name: feed.name,
+              type: 'unknown',
+              link: 'unknown'
+            }
+            fetcher = feed.fetcher_builder.call()
+            case fetcher
+            when Threatinator::Fetchers::Http
+              info[:type] = "http"
+              info[:link] = fetcher.url
+            end
+
+            feeds << info
+          end
+          feeds.sort! { |a,b| [a[:provider], a[:name]] <=> [b[:provider], b[:name]] }
+
+          io_out.write(MultiJson.dump(feeds))
+        end
+
+        def exec
+          case @config.format
+          when 'table'
+            output_table($stdout)
+          when 'json'
+            output_json($stdout)
+          else
+            raise ArgumentError, "Invalid argument for 'format' = '#{@config.format}'"
+          end
         end
       end
     end